- Convert BGRX texture to RGBA format in DDS image loader - Fix x64 build warnings.

+++ src/priv/engine/static_mesh.cpp
@@ -161,7 +161,7 @@ bool GN::engine::StaticMesh::loadFromFatModel( const GN::gfx::FatModel & fatmode
 
     DynaArray<uint8> vb;
 
-    for( size_t i = 0; i < fatmodel.meshes.size(); ++i )
+    for( uint32 i = 0; i < fatmodel.meshes.size(); ++i )
     {
         if( NULL == fatmodel.meshes[i] ) continue;
